Fixed several issues with the /go post function in main.js
Implemented a function to check a users progress through the lessons without use of a global variable. This function was then used and the global variable was removed. Also, made it so a user with a cookie will be directed to the farthest tutorial they have yet to complete when they click 'start here'. For example, a user who has completed lessons 1 and 2 will now be directed to Lesson 3 after clicking 'start here'